move communication card into own component
